Move CSF package to monorepo
