+++ code/core/src/csf/toStartCaseStr.ts
