+++ code/lib/cli/core/csf/index.cjs
@@ -1,0 +1,1 @@
+module.exports = require('@storybook/core/csf');
