+++ code/lib/cli/core/csf/index.d.ts
@@ -1,0 +1,2 @@
+export * from '@storybook/core/csf';
+export type * from '@storybook/core/csf';
