+++ code/lib/cli/core/csf/index.js
@@ -1,0 +1,1 @@
+export * from '@storybook/core/csf';
