+++ code/lib/cli/package.json
@@ -193,6 +193,11 @@
     },
     "./internal/preview/runtime": {
       "import": "./core/preview/runtime.js"
+    },
+    "./internal/csf": {
+      "types": "./core/csf/index.d.ts",
+      "import": "./core/csf/index.js",
+      "require": "./core/csf/index.cjs"
     }
   },
   "main": "dist/index.cjs",
@@ -232,6 +237,9 @@
       ],
       "internal/core-server": [
         "./core/core-server/index.d.ts"
+      ],
+      "internal/csf": [
+        "./core/csf/index.d.ts"
       ],
       "internal/csf-tools": [
         "./core/csf-tools/index.d.ts"
